backend/requirements.txt
fastapi 
uvicorn 
openai 
pydantic[dotenv] 
python-dotenv
google-generativeai
requests